# api/swagger/shop-api.yml
openapi: 3.0.0
info:
  title: SHOP App
  description: |
    API provides possibility to get products, products cart, orders
    and ability to authorization, sign in,  order products, reorder.
  version: 1.0.0
servers:
  - url: http://localhost:8000/api
    description: Local host for debugging
security:
  - bearerAuth: []
paths:
...
  /products/{product_id}/:
    get:
      summary: Get product by id
      parameters:
        - $ref: '#/components/parameters/product_id'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        404:
          description: Non-existent product id
  /products/:
    get:
      summary: Get paginated products
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/page_size'
        - $ref: '#/components/parameters/category_id_in_query'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedProducts'
  /categories/{category_id}/:
    get:
      parameters:
        - $ref: '#/components/parameters/category_id_in_path'
        - $ref: '#/components/parameters/include_products'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
  /categories/:
    get:
      summary: Get categories
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/page_size'
        - $ref: '#/components/parameters/include_products'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedCategories'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    page_size:
      name: page_size
      in: query
      schema:
        type: integer
        default: 5
    page:
      name: page
      in: query
      schema:
        type: integer
        default: 1
    product_id:
      name: product_id
      in: path
      required: true
      schema:
        type: integer
    category_id_in_path:
      name: category_id
      in: path
      schema:
        type: integer
    category_id_in_query:
      name: category_id
      in: query
      schema:
        type: integer
    include_products:
      name: include_products
      in: query
      schema:
        type: boolean
  schemas:
    Pagination:
      type: object
      properties:
        count:
          type: integer
        next:
          type: string
          nullable: true
        previous:
          type: string
          nullable: true
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        price:
          type: number
        category_id:
          type: integer
          example: 1
...
    PaginatedProducts:
      allOf:
        - $ref: '#/components/schemas/Pagination'
        - type: object
          properties:
            results:
              type: array
              items:
                $ref: '#/components/schemas/Product'
    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        products:
          type: array
          required: false
          items:
            $ref: '#/components/schemas/Product'
    PaginatedCategories:
      allOf:
        - $ref: '#/components/schemas/Pagination'
        - type: object
          properties:
            results:
              type: array
              items:
                $ref: '#/components/schemas/Category'
...